Refactor: Overhaul main codebase structure
