--- addr-book.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.8.34525.116
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "addr-book", "addr-book.vcxproj", "{6AC50086-82EF-45A7-AA17-C3CE54DB0B2C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6AC50086-82EF-45A7-AA17-C3CE54DB0B2C}.Debug|x64.ActiveCfg = Debug|x64
-		{6AC50086-82EF-45A7-AA17-C3CE54DB0B2C}.Debug|x64.Build.0 = Debug|x64
-		{6AC50086-82EF-45A7-AA17-C3CE54DB0B2C}.Debug|x86.ActiveCfg = Debug|Win32
-		{6AC50086-82EF-45A7-AA17-C3CE54DB0B2C}.Debug|x86.Build.0 = Debug|Win32
-		{6AC50086-82EF-45A7-AA17-C3CE54DB0B2C}.Release|x64.ActiveCfg = Release|x64
-		{6AC50086-82EF-45A7-AA17-C3CE54DB0B2C}.Release|x64.Build.0 = Release|x64
-		{6AC50086-82EF-45A7-AA17-C3CE54DB0B2C}.Release|x86.ActiveCfg = Release|Win32
-		{6AC50086-82EF-45A7-AA17-C3CE54DB0B2C}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FF637A6F-9265-4FFE-A670-3ECC8AF3CEE7}
-	EndGlobalSection
-EndGlobal
